decent state for states lol
